vercel-telegram-docx/requirements.txt
python-telegram-bot==21.4
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-docx==1.1.2
